+++ src/main/kotlin/com/autonomousapps/extension/UsageAnalysisHandler.kt
@@ -12,6 +12,9 @@ import javax.inject.Inject
  * dependencyAnalysis {
  *   usage {
  *     analysis {
+ *       // When true, does binary compatibility analysis, which can be very expensive in memory terms.
+ *       checkBinaryCompatibility(true) // defaults to false
+ *
  *       // When true, do superclass analysis to ensure necessary superclasses are on the classpath. This analysis is
  *       // very expensive.
  *       checkSuperClasses(true) // defaults to false
@@ -22,9 +25,14 @@ import javax.inject.Inject
  */
 public abstract class UsageAnalysisHandler @Inject constructor(objects: ObjectFactory) {
 
+  internal val checkBinaryCompat = objects.property(Boolean::class.java).convention(false)
   internal val checkSuperClasses = objects.property(Boolean::class.java).convention(false)
 
-  @Suppress("unused") // public API
+  public fun checkBinaryCompatibility(shouldCheck: Boolean) {
+    checkBinaryCompat.set(shouldCheck)
+    checkBinaryCompat.disallowChanges()
+  }
+
   public fun checkSuperClasses(shouldCheck: Boolean) {
     checkSuperClasses.set(shouldCheck)
     checkSuperClasses.disallowChanges()